refactor(api-server): split routes + extract middleware + Upstash ratelimit + Vercel waitUntil
Pulled inspiration from dub.co and midday-ai for structure:
- Per-resource sub-routers under apps/api-server/src/routes/{cohorts,students,persons}.ts
- Reusable middleware extracted into src/middleware/{auth,scope,rate-limit,audit-log,body-capture,request-id}.ts
- withRequiredScope as a per-route createRoute middleware (short-circuits before validation)
- Replaced in-process token-bucket with Upstash Ratelimit (sliding window 60/min)
- Background audit-log writes via @vercel/functions waitUntil — no longer blocks the response
- Scalar API reference UI at /docs, OpenAPI doc at /v1/openapi.json
- Switched auth.ts drizzle helpers to @nawadi/db/drizzle to dedupe drizzle install (avoids the @upstash/redis optional peer fork)
