+++ apps/api-server/package.json
@@ -36,7 +36,11 @@
     "@hono/zod-openapi": "^1.3.0",
     "@nawadi/core": "workspace:*",
     "@nawadi/db": "workspace:*",
+    "@scalar/hono-api-reference": "^0.5.180",
     "@types/node": "catalog:",
+    "@upstash/ratelimit": "^2.0.6",
+    "@upstash/redis": "^1.35.5",
+    "@vercel/functions": "^3.4.4",
     "better-auth": "^1.6.9",
     "drizzle-orm": "catalog:",
     "hono": "^4.6.14",
