+++ apps/api-server/src/middleware/index.ts
@@ -1,0 +1,6 @@
+export { requireAuth } from "./auth.js";
+export { auditLog } from "./audit-log.js";
+export { bodyCapture, isPiiRedactedPath } from "./body-capture.js";
+export { rateLimit } from "./rate-limit.js";
+export { requestId } from "./request-id.js";
+export { withRequiredScope } from "./scope.js";
