+++ apps/api-server/src/types.ts
@@ -1,0 +1,10 @@
+import type { ApiClient } from "./auth.js";
+
+export type Variables = {
+  client: ApiClient;
+  requestId: string;
+  startedAt: number;
+  requestBody: unknown;
+};
+
+export type Env = { Variables: Variables };
